SERVER-70000 pass literal suffices explicitly to BOOST_MP_DEFINE_SIZED_CPP_INT_LITERAL (#1)
In our testing environment MSVC 14.31.31103 emits a compile error in two instances involving the concatenation of a string literal with a macro argument and the use of the concatenation result as the name of a user-defined literal. This commit solves the compile error by passing the precomputed results of the concatenations to the macro.